refactor helper, so its visible to client package
